renamed scanner page
