+++ apps/app/src/pages/ScannerPage/ScannerPage.module.scss
