+++ apps/app/src/pages/ScannerPage/index.ts
