+++ apps/app/src/pages/ScannerPage/popups/ScannedAchievementPopup.module.scss
