+++ apps/app/src/pages/ScannerPage/popups/ScannedAchievementPopup.tsx
